Apply new rules of attribute key:
Attribute key can only match 2 formats:
- String ("data-key"="value")
- Symbol with hyphen (data-key="value")
