+++ lib/Tag.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const assert = require('assert');
-const parseSignature = require('./parseSignature')
+const parseAttributes = require('./parseAttributes')
 
 /**
  * 自定义标签的基类，封装 tag 的通用逻辑。
@@ -60,7 +60,7 @@ class Tag {
 
     // parse the args and move after the block end. passing true
     // as the second arg is required if there are no parentheses
-    let args = parseSignature.call(parser, null, true);
+    let args = parseAttributes.call(parser);
     parser.advanceAfterBlockEnd(token.value);
 
     let body;
